Correccion a Inventarios Fisicos

+++ ERPAPI/Models/Inventarios/InventarioFisicoLine.cs
@@ -38,7 +38,7 @@ namespace ERPAPI.Models
 
         public decimal? FactorSacos { get; set; }
 
-        public int? NSacos { get; set; }
+        public decimal? NSacos { get; set; }
 
         public decimal SaldoLibros { get; set; }
